decompose nba matchup syncer into base class
